fixed env

+++ backend/controllers/productControllers.js
@@ -51,6 +51,9 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+		return res.status(404).json({ success: false, message: "Invalid Product Id" });
+	}
     
     try {
         const deleteProduct = await Product.findByIdAndDelete(id);
